perf: Do not use smaller uint unnecessarily (#170)
* perf: Do not use smaller uint unnecessarily

* perf: Do not use smaller uint unnecessarily (v2)

+++ contracts/interfaces/ILooksRareProtocol.sol
@@ -20,7 +20,7 @@ interface ILooksRareProtocol {
 
     struct SignatureParameters {
         bytes32 orderHash;
-        uint128 orderNonce;
+        uint256 orderNonce;
         bool isNonceInvalidated;
         address signer;
     }
